requirements.txt
argparse==1.2.1
boto==2.36.0
django-angular==0.7.13
django-bootstrap-form==3.1
django-debug-toolbar==1.3.0
django-js-reverse==0.3.3
sqlparse==0.1.15
wsgiref==0.1.2


# Authwrap written by jamie-w

# git repo
-e git+https://bitbucket.org/dukeofweatherby/django-authwrap.git#egg=django-authwrap-authwrap-mixins

# local version
# -e /Users/jamie/Code/home/django-authwrap 
